Add mutt rc
For the password: to avoid having to type it everytime, but at the same time
not having to store it into this rc as plain text (ugh), write a text file
with the content

    set my_pass = "<password>"

then PGP-encrypt it into .my-pwds.gpg, and have mutt decrypt it into a local
variable to use at login. This will only spawn a password prompt for my GPG
key, and not even that if I have a GPG agent.