Added functions to nvim to link to default server socket
Neovim automatically creates a socket that can be used for --remote-send
(install neovim-remote via pip and use the nvr program). However, in
order to not have to explicitly specify the socket (which is a numbered
file in a random-name folder in /tmp), I wrote a function to make a
symlink to nvr's default socket file, /tmp/nvimsocket, and one to clear
the symlink, which is automatically done when closing the instance.

Also changed the ZSH g alias to point to nvr. This of course requires
neovim-remote to be installed.