Added method to switch mpd hosts for local controls
Needs a file in ~/.config/mpd called mpdhost which may contain a
commented out and a non-commented out export statement defining the
MPD_HOST. This file is sourced by mpcctl and polybar/mpd and makes it
super easy to switch the host (raspi/local).
Also added an i3 shortcut to switch host.